Tests fixed for: Spec intention/spec fix
