+++ testData/quickfixes/generate_spec/withArgs-after.erl
@@ -1,2 +1,2 @@
--spec foo(any(), any(), any()) -> atom().
+-spec foo(One :: any(), Two :: any(), Three :: any()) -> atom().
 foo(One, Two, Three) -> ok.
